% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_delimitted_param.R
\name{parse_delimitted_param}
\alias{parse_delimitted_param}
\title{Function to parse delimitted parameter}
\usage{
parse_delimitted_param(param_str, delim = ",", rm.empty = T)
}
\arguments{
\item{param_str}{character. String to be parsed.}

\item{delim}{character. The delimiter.}

\item{rm.empty}{logical. If TRUE, removes every empty string after parsing.}
}
\value{
A vector of parsed character strings.
}
\description{
This function splits the given string by a delimiter. It can also remove empty strings.
}
\examples{
x = "one,two,three,"
parsed_strings = parse_delimitted_param(x, delim = ',', rm.empty = T)
}
